Refactor status command
